Revert to single python file job
This commit removes setup.py and git submodules. Instead of using
setup.py to install the package in the ATMO cluster, we're going to
download the repo and just run the main.py file.

We were trying to install the job as a package so we could automatically
install the dependencies, simplifying the setup. However, when we build
an egg file to include in the spark config, the dependencies aren't
included. This means we have to install dependencies seperately.